+++ CHANGELOG.md
@@ -6,10 +6,11 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
-## 0.30.1 (TBD)
+## 0.31.0 (TBD)
 
 ### new:
-- The `bert_model` argument in the `text.ner.models.sequence_tagger` function now supports models like Roberta (for `bio-bert` NER models).
+- The `text.ner.models.sequence_tagger` now supports embeddings from non-BERT transformer models (e.g., `roberta-base`, `openai-gpt`)
+- [**breaking change**] In the `text.ner.models.sequence_tagger` function, the `bilstm-bert` model  is now called `bilstm-transformer` and the `bert_model` parameter has been renamed to `transormer_model`. 
 
 ### changed
 - Code base was reformatted using [black](https://github.com/psf/black)